requirements.txt
asgiref==3.10.0
attrs==25.4.0
black==25.9.0
cffi==2.0.0
cfgv==3.4.0
click==8.3.0
coverage==7.11.3
cryptography==46.0.3
distlib==0.4.0
Django==5.2.7
django-ipware==7.0.1
django-structlog==10.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==38.0.0
filelock==3.20.0
gunicorn==23.0.0
identify==2.6.15
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-decouple==3.8
python-ipware==3.0.0
python-json-logger==4.0.0
pytokens==0.2.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.14.3
sqlparse==0.5.3
structlog==25.5.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.35.4
